Projectiles from the cannon shoot the same angle on client and server but the cannon texture does not face the same angle.
Added cannon angle to packets to be sent across network so which ever way the client shoots the projectile will shoot that way on the server. Only problem is that the cannon texture is a bit off the angle more and more when you move. It won't effect game play but looks bad to someone watching the other player fire.